% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_labeller.R
\name{censo_labeller}
\alias{censo_labeller}
\title{Label microdata variables 2010 Census persons}
\usage{
censo_labeller(data_censo, dictionary.file)
}
\arguments{
\item{data_censo}{An object of the 2010 Census persons microdata read with
\code{read_censo} function.}

\item{dictionary.file}{The dictionary file for selected survey available on
official website: \url{https://ftp.ibge.gov.br/Censos/Censo_Demografico_2010/Resultados_Gerais_da_Amostra/Microdados/}}
}
\description{
This function will label the microdata variables refering to
available information about the individuals of each of the households
available in the 2010 Census.
}
