% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_labeller.R
\name{census_labeller}
\alias{census_labeller}
\title{Label microdata variables 2010 Census persons}
\usage{
census_labeller(data_census, dictionary.file)
}
\arguments{
\item{data_census}{An object of the 2010 Census persons microdata read with
\code{read_censo} function.}

\item{dictionary.file}{The dictionary file for selected survey available on
official website: \url{https://ftp.ibge.gov.br/Censos/Censo_Demografico_2010/Resultados_Gerais_da_Amostra/Microdados/}}
}
\description{
This function will label the microdata variables refering to
available information about the individuals of each of the households
available in the 2010 Census.
}
