% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census.R
\name{get_census}
\alias{get_census}
\title{Download and label for microdata persons 2010 Census}
\usage{
get_census(state, vars = NULL, savedir = tempdir())
}
\arguments{
\item{state}{Vector of the states for which you want to obtain the microdata}

\item{vars}{Vector of variable names to be kept for analysis. Default is to
kept all variables}

\item{savedir}{Directory to save the download data. Default is to use a
temporary directory}
}
\description{
This function will download and label the microdata refering
to the available information about the individuals of each of the households
available in the 2010 Census.
}
