% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_censo.R
\name{read_censo}
\alias{read_censo}
\title{Read microdata persons 2010 Census}
\usage{
read_censo(microdata, input_txt, vars = NULL)
}
\arguments{
\item{microdata}{A text file containing microdata from 2010 Census persons
survey, available on official website: \url{https://ftp.ibge.gov.br/Censos/Censo_Demografico_2010/Resultados_Gerais_da_Amostra/Microdados/}}

\item{input_txt}{A text file, related to the microdata from 2010 Census
persons, containing the input script for SAS, available on the official
website: \url{https://ftp.ibge.gov.br/Censos/Censo_Demografico_2010/Resultados_Gerais_da_Amostra/Microdados/}}

\item{vars}{Vector of variable names to be kept for analysis. Default is to
keep all variables.}
}
\description{
This function will read the microdata refering to the
information available on the individuals of each of the households
available in the 2010 Census.
}
